Fix for conflicting react versions.
Material UI was including React at 15.4.2, but the
application's React version was set to 15.3.2.  This was
causing all of the tests to fail due to multiple included
versions of React.

This change puts all of the versions into harmony.  At least,
they are harmonious enough for all of the tests to pass in master.