Remove artifacts from map
When the source is updated we need to remove the old features from the
draw instance, otherwise there will be artifacts in mapbox draw with a
layername `.cold` or similar. ensure we delete all the features from
draw before adding them again. this avoids adding features multiple
times as well
will add delete feature to the demo shortly after this PR.